so all that before? yeah forget that added CreateDefaultUser method to RoleHelper class and deleted IdentityHelper class since it was redundant.
