+++ LeagueScoreWebsite/Models/RoleHelper.cs
@@ -23,5 +23,28 @@ namespace LeagueScoreWebsite.Models
                 }
             }
         }
+
+        public static async Task CreateDefaultUser(IServiceProvider provider, string role)
+        {
+            var userManager = provider.GetService<UserManager<IdentityUser>>();
+
+            // if no users are present make default user
+            int numUsers = (await userManager.GetUsersInRoleAsync(role)).Count;
+            if (numUsers == 0) // if no users are in the specified role
+            {
+                var defaultUser = new IdentityUser()
+                {
+                    Email = "admin@leaguescoresite.com",
+                    // UserName = "Admin"
+                };
+
+                await userManager.CreateAsync(defaultUser, "thisisTHEadminPassw0rd!");
+
+                await userManager.AddToRoleAsync(defaultUser, role);
+            }
+        }
+
     }
+
+
 }
